[math440] Add notes for week 4.
I'm quite behind on the proofs. As an overview, though, we show Theorem
1.4 by considering the set of triples of the form (L, L', s) where L,L'
are isomorphic field extensions of F and s is the isomorphism between
them. We then bound all chains, and use Zorn's Lemma to say there is a
maximal, and use s to exhibit an isomorphism from L to K and L' to K'.

For Theorem 1.5, consider the lexicographically first monomial and, by
symmetry, it must be descending in exponents. This way, we can write
that monomial in terms of elementary symmetric polynomials (perhaps with
leftover) and repeat this process by way of infinite descent (we argue
that the leading monomial is lexicographically increasing through this
process.

For Theorem 1.6, we proceed by induction on m for polynomials of degree
k2^m with odd k. We use the symmetric polynomial for each lambda in R:
\[p_\lambda = \prod_{i<j}(x-(a_i + a_j) + \lambda a_i a_j)\]
for roots a_i going over all the roots of f.